+++ src/bitmap/binary_write.hpp
