+++ src/bitmap/bitmap.hpp
