+++ src/bitmap/bitmap_io.hpp
