+++ src/bitmap/detail/binary_io_flags.hpp
