+++ src/bitmap/detail/io.hpp
