+++ src/bitmap/detail/valid_binary_format.hpp
