+++ src/bitmap/exception.hpp
