+++ src/bitmap/get_size.hpp
