+++ src/bitmap/histogram.hpp
