+++ src/bitmap/interpolate.hpp
