+++ src/bitmap/masked_pixel.hpp
