+++ src/bitmap/matrix3x3.hpp
