+++ src/bitmap/pixel.hpp
