+++ src/bitmap/pixel_algorithm.hpp
