+++ src/bitmap/pixel_output.hpp
