+++ src/bitmap/point.hpp
