+++ src/bitmap/point_io.hpp
