+++ src/bitmap/rect.hpp
