+++ src/bitmap/rect_io.hpp
