+++ src/bitmap/rect_transform.hpp
