+++ src/bitmap/size.hpp
