+++ src/bitmap/size_io.hpp
