+++ src/bitmap/subbitmap.hpp
