+++ src/bitmap/transform.hpp
